Add failing test (plus some minor changes)
